Use intptr_t and uintptr_t for integer/pointer conversions
Use intptr_t and uintptr_t casts to appease the compiler when
values are converted between 32-bit integers and pointers.
This generates many warnings (currently ~170 on linux amd64)
when pointers are 64-bit.

These types are normally defined in stdint.h, and get defined
by autoconf if that's not the case.

In a few places, NULL is simply replaced with 0 where compared
with an integer.

Reviewed-on: http://gerrit.openafs.org/474
Tested-by: Derrick Brashear <shadow@dementia.org>

+++ src/afsmonitor/afsmonitor.c
@@ -3768,7 +3768,7 @@ afsmon_execute(void)
     }
 
     /* start the gtx input server */
-    code = (int) gtx_InputServer(afsmon_win);
+    code = (intptr_t)gtx_InputServer(afsmon_win);
     if (code) {
 	fprintf(stderr, "[ %s ] Failed to start input server \n", rn);
 	afsmon_Exit(140);
